add login/signup flow

+++ src/routes/api/projects/+server.js
@@ -3,6 +3,9 @@ import { json } from '@sveltejs/kit';
 
 /** @type {import('./$types').RequestHandler} */
 export async function DELETE({ request }) {
+  if (!locals.session.data.userId) {
+    throw redirect(302, '/users/login');
+  }
   const { projectId } = await request.json();
   await deleteProject(+projectId);
