+++ src/routes/api/tasks/+server.js
@@ -1,10 +1,13 @@
 import { deleteTask } from '$lib/server/db';
-import { json } from '@sveltejs/kit';
+import { json, redirect } from '@sveltejs/kit';
 
 /** @type {import('./$types').RequestHandler} */
-export async function DELETE({ request }) {
+export async function DELETE({ request, locals }) {
+  if (!locals.session.data.userId) {
+    throw redirect(302, '/users/login');
+  }
   const { taskId } = await request.json();
   const task = await deleteTask(+taskId);
-  
+
   return json({ task });
 }
